+++ src/languages/dotnet/dotnetUtility.ts
@@ -1,17 +1,17 @@
 "use strict";
-import * as path from "path";
 import * as vscode from "vscode";
+import { Constants } from "../../common/constants";
 import { Executor } from "../../common/executor";
 import { TelemetryClient } from "../../common/telemetryClient";
 import { Utility } from "../../common/utility";
 
 export class DotnetUtility {
-    public static dotnetPublish(fileUri: vscode.Uri) {
-        TelemetryClient.sendEvent("dotnetPublish.start");
-        if (!fileUri) {
-            return;
+    public async dotnetPublish(fileUri?: vscode.Uri) {
+        const projectFilePath: string = await Utility.getInputFilePath(fileUri, Constants.dotNetProjectFileNamePattern, ".NET project file", "dotnetPublish.start");
+
+        if (projectFilePath) {
+            Executor.runInTerminal(`dotnet publish \"${Utility.adjustFilePath(projectFilePath)}\"`);
+            TelemetryClient.sendEvent("dotnetPublish.end");
         }
-        Executor.runInTerminal(`dotnet publish \"${Utility.adjustFilePath(fileUri.fsPath)}\"`);
-        TelemetryClient.sendEvent("dotnetPublish.end");
     }
 }